split up context.rs
